impr group callbacks, remove legacy
